# Improve performance of identifying jump items