--- a/src/engraving/dom/chordrest.cpp
+++ b/src/engraving/dom/chordrest.cpp
@@ -1302,62 +1302,19 @@ bool ChordRest::hasFollowingJumpItem() const
     if (!measure) {
         return false;
     }
-    const Fraction nextTick = seg->tick() + actualTicks();
 
     if (measure->lastChordRest(track()) != this) {
         return false;
     }
 
-    // Jumps & markers
-    for (const EngravingItem* e : measure->el()) {
-        if (!e->isJump() && !e->isMarker()) {
-            continue;
-        }
+    std::vector<Measure*> followingRepeatMeasures = findFollowingRepeatMeasures(measure);
 
-        if (e->isJump()) {
-            return true;
-        }
-
-        const Marker* marker = toMarker(e);
-
-        if (marker->isRightMarker()) {
-            return true;
-        }
-    }
-
-    // Voltas
-    auto spanners = score()->spannerMap().findOverlapping(measure->endTick().ticks(), measure->endTick().ticks());
-    for (auto& spanner : spanners) {
-        if (!spanner.value->isVolta() || Fraction::fromTicks(spanner.start) != nextTick) {
-            continue;
-        }
-
-        return true;
-    }
-
-    // Repeats
-    if (measure->repeatEnd()) {
-        return true;
-    }
-
-    for (Segment* nextSeg = seg->next(SegmentType::BarLineType); nextSeg && nextSeg->tick() == nextTick;
-         nextSeg = nextSeg->next(SegmentType::BarLineType)) {
-        const EngravingItem* el = nextSeg->element(track());
-        if (!el || !el->isBarLine()) {
-            continue;
-        }
-        const BarLine* bl = toBarLine(el);
-
-        if (bl->barLineType() & (BarLineType::END_REPEAT | BarLineType::END_START_REPEAT)) {
-            return true;
-        }
-    }
-
-    return false;
+    return !followingRepeatMeasures.empty();
 }
 
 bool ChordRest::hasPrecedingJumpItem() const
 {
+    TRACEFUNC;
     const Segment* seg = segment();
     const Measure* measure = seg->measure();
 
@@ -1365,52 +1322,8 @@ bool ChordRest::hasPrecedingJumpItem() const
         return false;
     }
 
-    if (seg->score()->firstSegment(SegmentType::ChordRest) == seg) {
-        return true;
-    }
+    std::vector<Measure*> precedingRepeatMeasures = findPreviousRepeatMeasures(measure);
 
-    // Markers
-    for (const EngravingItem* e : measure->el()) {
-        if (!e->isMarker()) {
-            continue;
-        }
-
-        const Marker* marker = toMarker(e);
-        if (marker->isRightMarker()) {
-            continue;
-        }
-
-        return true;
-    }
-
-    // Voltas
-    auto spanners = score()->spannerMap().findOverlapping(measure->tick().ticks(), measure->tick().ticks());
-    for (auto& spanner : spanners) {
-        if (!spanner.value->isVolta() || Fraction::fromTicks(spanner.start) != tick() || toVolta(spanner.value)->isFirstVolta()) {
-            continue;
-        }
-
-        return true;
-    }
-
-    // Repeat barlines
-    if (measure->repeatStart()) {
-        return true;
-    }
-
-    for (Segment* prevSeg = seg->prev(SegmentType::BarLineType); prevSeg && prevSeg->tick() == seg->tick();
-         prevSeg = prevSeg->prev(SegmentType::BarLineType)) {
-        EngravingItem* el = prevSeg->element(track());
-        if (!el || !el->isBarLine()) {
-            continue;
-        }
-
-        BarLine* bl = toBarLine(el);
-        if (bl->barLineType() & (BarLineType::START_REPEAT | BarLineType::END_START_REPEAT)) {
-            return true;
-        }
-    }
-
-    return false;
+    return !precedingRepeatMeasures.empty();
 }
 }