This PR replaces a more complex set of checks to see if a `ChordRest` is followed by a repeat item with simply checking against the score's repeat structure.
I was trying to avoid forcing the score to update its repeat structure more often than was necessary. What I hadn't factored in was updating the score's spanner map. Calls to `SpannerMap::findOverlapping` were updating the whole spanner map very frequently, leading to a large impact in performance. 
While the new solution does regenerate the score's repeat structure more often than before, this is a much faster operation - there will usually be fewer repeat sections than spanners in a score. Testing on the Beethoven 9 score in release mode:

|   | NotationProject::load(ms) | Score::doLayoutRange(ms) | hasFollowingJumpItem(ms) | hasPrecedingJumpItem(ms) |
| ------------- | ------------- | ------------- | ------------- | ------------- | 
| Before PR 1  | 40990.274 | 28402.757 | 7253.991 | 366.372 |
| Before PR 2 | 40205.233  | 28287.846 | 6874.989 | 370.329 | 
| After PR 1 | 36470.430 | 28058.989 | 1851.505 | 14.313 | 
| After PR 2 | 36806.955 | 28424.238 | 1815.317 | 16.000 | 


The second commit makes similar changes to `segmentsAreAdjacentInRepeatStructure`. The performance improvements for this measure are as follows:
`Before PR 1   0.769 ms    5491    4220.364 ms`
`Before PR 2   0.795 ms    5491    4367.959 ms`
`After PR  1   0.001 ms    5491    2.873 ms`
`After PR  2   0.002 ms    5491    10.280 ms`

Also:
Resolves: #27303